refactor: screen 상수 값을 추가합니다.
