+++ src/constants/onboarding.ts
@@ -1,0 +1,26 @@
+export const ON_BOARDING = [
+  {
+    title: '안녕하세요',
+    heading: `위프로는 효과적으로
+자신을 어필할 수 있도록
+돕는 서비스입니다`,
+  },
+  {
+    title: '무슨 방법으로요?',
+    heading: `프로젝트에 따라
+서로의 기술과 협업 능력을
+평가할 수 있어요`,
+  },
+  {
+    title: '어떻게 보여주나요?',
+    heading: `팀원의 평가와 함께
+나를 표현할 수 있는
+명함을 받아요`,
+  },
+  {
+    title: '누가 답변했는지 알 수 있나요?',
+    heading: `프로젝트에 참여한 사실만
+노출되고 어떤 답변을 했는지
+알 수 없어요`,
+  },
+] as const;
