+++ src/constants/service.ts
@@ -1,20 +1,3 @@
-export const ON_BOARDING = [
-  {
-    title: '안녕하세요',
-    heading: `위프로는 효과적으로
-자신을 어필할 수 있도록
-돕는 서비스입니다`,
-  },
-  {
-    title: '무슨 방법으로요?',
-    heading: `프로젝트에 따라
-서로의 기술과 협업 능력을
-평가할 수 있어요`,
-  },
-  {
-    title: '어떻게 보여주나요?',
-    heading: `팀원의 평가와 함께
-나를 표현할 수 있는
-명함을 받아요`,
-  },
-] as const;
+export const SCREEN_SIZE = {
+  Web: 375,
+} as const;
